refactor(logic): move view and function interfaces from provider to contracts instances
