+++ .prettierrc
@@ -5,5 +5,8 @@
   "tabWidth": 2,
   "singleQuote": true,
   "semi": false,
-  "arrowParens": "avoid"
+  "arrowParens": "avoid",
+  "importOrder": ["^[./]"],
+  "importOrderSeparation": true,
+  "importOrderSortSpecifiers": true
 }
