+++ src/assisted-provider.ts
@@ -4,15 +4,7 @@ import { BaseProvider } from './base-provider'
 import { Pools as PoolsClass } from './pools'
 import { ERC20CollateralPool } from './erc20-collateral-pool'
 import { miscErc20CollateralPool } from './artifacts'
-import {
-  Pool as Erc20CollateralPool,
-  Borrow,
-  Lend
-} from './types/erc20-collateral-token'
-import { Pool as Pools, PoolCommit } from './types/pools'
 import { BaseContractConstructorParams } from './base-contract'
-
-type Erc20CollateralTokenPoolDetail = Borrow | Lend
 
 export { Abi, PrivateKey, Pool }
 
@@ -30,35 +22,5 @@ export class AssistedProvider<
     super(new contractBuilder(address, apiUrl, null, abi))
 
     this.abi = abi || miscErc20CollateralPool.abi
-  }
-
-  getPool(poolId: bigint): Promise<Erc20CollateralPool | Pools> {
-    throw new Error(`Method not implemented. ${poolId.toString()}`)
-  }
-
-  getPools(
-    offset: bigint,
-    limit: bigint
-  ): Promise<Array<Erc20CollateralPool | Pools>> {
-    throw new Error(
-      `Method not implemented. ${offset.toString()}, ${limit.toString()}`
-    )
-  }
-
-  getPoolDetails(
-    poolId: bigint,
-    walletAddress: string
-  ): Promise<Erc20CollateralTokenPoolDetail | PoolCommit> {
-    throw new Error(
-      `Method not implemented. ${poolId.toString()}, ${walletAddress}`
-    )
-  }
-
-  pause(): Promise<void> {
-    throw new Error('Method not implemented.')
-  }
-
-  unpause(): Promise<void> {
-    throw new Error('Method not implemented.')
   }
 }
