+++ src/pools.ts
@@ -1,5 +1,7 @@
 import { Abi, PrivateKey } from './types/types'
-import { BaseContract } from './base-contract'
+import { BaseContract, Erc20CollateralTokenPoolDetail } from './base-contract'
+import { Pool as Erc20CollateralPool } from './types/erc20-collateral-token'
+import { Pool as PoolsObject, PoolCommit } from './types/pools'
 
 export class Pools extends BaseContract {
   constructor(
@@ -9,5 +11,35 @@ export class Pools extends BaseContract {
     abi?: Abi
   ) {
     super(address, apiUrl, privateKey, abi)
+  }
+
+  getPool(poolId: bigint): Promise<Erc20CollateralPool | PoolsObject> {
+    throw new Error(`Method not implemented. ${poolId.toString()}`)
+  }
+
+  getPools(
+    offset: bigint,
+    limit: bigint
+  ): Promise<Array<Erc20CollateralPool | PoolsObject>> {
+    throw new Error(
+      `Method not implemented. ${offset.toString()}, ${limit.toString()}`
+    )
+  }
+
+  getPoolDetails(
+    poolId: bigint,
+    walletAddress: string
+  ): Promise<Erc20CollateralTokenPoolDetail | PoolCommit> {
+    throw new Error(
+      `Method not implemented. ${poolId.toString()}, ${walletAddress}`
+    )
+  }
+
+  pause(): Promise<void> {
+    throw new Error('Method not implemented.')
+  }
+
+  unpause(): Promise<void> {
+    throw new Error('Method not implemented.')
   }
 }
