+++ src/provider-utilities.ts
@@ -1,5 +1,3 @@
 export const Role = {
   ADMIN: '0x' + '00'.repeat(32)
 } as const
-
-export class Provider {}
